Implement prevention of same field names in streams and channel properties table
If some of the fields in input stream feature class is the same name as any of the fields in channel properties table it is removed during the clipping to AoI and user is informed.
The _stream_clip method renamed to _clip_streams to match other clipping methods in the data preparation.